edit component
